@@ -1,7 +1,7 @@
 [package]
 name = "private_speech"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 authors = ["Chase C <hi@chasecares.dev>"]
 license = "GPL-2.0-only"
 
