@@ -1,3 +1,5 @@
+use ksni::menu::StandardItem;
+
 pub struct Menu {
     pub playing: bool,
     pub status: String,
@@ -20,13 +22,14 @@ impl ksni::Tray for Menu {
     }
 
     fn menu(&self) -> Vec<ksni::MenuItem<Self>> {
-        use ksni::menu::*;
-        vec![StandardItem {
-            label: "Exit".into(),
-            icon_name: "application-exit".into(),
-            activate: Box::new(|_| std::process::exit(0)),
-            ..Default::default()
-        }
-        .into()]
+        vec![
+            StandardItem {
+                label: "Exit".into(),
+                icon_name: "application-exit".into(),
+                activate: Box::new(|_| std::process::exit(0)),
+                ..Default::default()
+            }
+            .into(),
+        ]
     }
 }
